requirements.txt
Flask==3.1.1
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.0
sympy==1.12
sentence-transformers==3.0.1
faiss-cpu==1.8.0
langchain==0.3.1
langchain-community==0.3.1
langchain-huggingface==0.2.1
torch==2.3.1
python-dotenv==1.0.1
requests==2.32.3
openpyxl==3.1.5